Add new home and splash views with animations
Introduce new home and changed splash views then added animated brand logo and transition
animations. Implement sliding and fading animations for splash view.

implemented a method to navigate to home with transition animation
