+++ lib/Features/Home/Presentation/View/home_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HomeView extends StatelessWidget {
+  const HomeView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+          child: Text(
+        'Home Screen',
+      )),
+    );
+  }
+}
