+++ lib/Features/Splash/Presentation/Views/splash_view.dart
@@ -1,4 +1,4 @@
-import 'package:book_store/Features/splash/Presenation/Views/Widgets/splash_view_body.dart';
+import 'package:book_store/Features/Splash/Presentation/Views/Widgets/splash_view_body.dart';
 import 'package:flutter/material.dart';
 
 class SplashView extends StatelessWidget {
@@ -7,7 +7,7 @@ class SplashView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      body: SplasgViewBody(),
+      body: SplashViewBody(),
     );
   }
 }
