+++ lib/constants.dart
@@ -1,3 +1,5 @@
 import 'package:flutter/material.dart';
 
 const kPrimaryColor = Color(0xFF1E1E1E);
+
+const kTransitionDuration = Duration(milliseconds: 500);
